moved RegExGenerator Interface into package
